Consolidate server implementations: merge serve modules and unify configuration
- Merge serve_with_client_manager.rs into serve.rs (277 lines removed)
- Extend Config with optional client management fields
- Add serve_with_client_management() that returns (ClientManager, Future)
- Remove implicit spawning - users control server execution
- Maintain backward compatibility with serve() function
- Update dynamic_client_management example to use unified API
